requirements.txt
asgiref==3.7.2
asttokens==2.4.1
build==1.0.3
CacheControl==0.13.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cleo==2.1.0
colorama==0.4.6
crashtest==0.4.1
cryptography==41.0.4
decorator==5.1.1
distlib==0.3.7
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
dulwich==0.21.6
executing==2.0.1
fastjsonschema==2.18.1
filelock==3.12.2
idna==3.4
importlib-metadata==6.8.0
inflection==0.5.1
installer==0.7.0
ipython==8.17.2
jaraco.classes==3.3.0
jedi==0.19.1
jeepney==0.8.0
keyring==24.2.0
matplotlib-inline==0.1.6
more-itertools==10.1.0
msgpack==1.0.7
packaging==23.2
parso==0.8.3
pexpect==4.8.0
Pillow==10.1.0
pipenv==2023.11.14
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.7.0
poetry-core==1.8.1
poetry-plugin-export==1.6.0
prompt-toolkit==3.0.41
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
pyproject_hooks==1.0.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.5.2
requests==2.31.0
requests-toolbelt==1.0.0
SecretStorage==3.3.3
shell==1.0.1
shellingham==1.5.3
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tomlkit==0.12.1
traitlets==5.13.0
trove-classifiers==2023.9.19
uritemplate==4.1.1
urllib3==2.0.6
virtualenv==20.24.6
wcwidth==0.2.10
zipp==3.17.0
